requirements.txt
langchain
openai
PyPDF2
faiss-cpu
huggingface-hub
python-dotenv
sentence-transformers
htmltemplate
